Add unique device naming #8
Since many devices can be in the same room, they require a unique ID. The
naming method splits at the trailing zeros of the UUID and takes the 5
characters immediately to the left of the split point. For example:

UUID: 7C 2C 67 CA 61 20 00 00 00 00 00 00
Split: 7C 2C 67 CA 61 2
Final Name: LECTEC-CA612

This naming convention serves as the device identifier for easy
identification in rooms with many skateboards.